+++ tests/TestCase.php
@@ -1,37 +1,24 @@
 <?php
 
-namespace VendorName\Skeleton\Tests;
+namespace Orlyapps\OrlyErrorTracking\Tests;
 
-use Illuminate\Database\Eloquent\Factories\Factory;
 use Orchestra\Testbench\TestCase as Orchestra;
-use VendorName\Skeleton\SkeletonServiceProvider;
+use Orlyapps\OrlyErrorTracking\OrlyErrorTrackingServiceProvider;
 
 class TestCase extends Orchestra
 {
-    protected function setUp(): void
-    {
-        parent::setUp();
-
-        Factory::guessFactoryNamesUsing(
-            fn (string $modelName) => 'VendorName\\Skeleton\\Database\\Factories\\'.class_basename($modelName).'Factory'
-        );
-    }
-
     protected function getPackageProviders($app)
     {
         return [
-            SkeletonServiceProvider::class,
+            OrlyErrorTrackingServiceProvider::class,
         ];
     }
 
     public function getEnvironmentSetUp($app)
     {
-        config()->set('database.default', 'testing');
-
-        /*
-         foreach (\Illuminate\Support\Facades\File::allFiles(__DIR__ . '/../database/migrations') as $migration) {
-            (include $migration->getRealPath())->up();
-         }
-         */
+        config()->set('cache.default', 'array');
+        config()->set('orly-error-tracking.enabled', true);
+        config()->set('orly-error-tracking.url', 'https://orly.test/api/error-tracking/v1/events');
+        config()->set('orly-error-tracking.key', 'project-key');
     }
 }